feat: ✨ add screen navigation system
