+++ app/src/main/java/org/gulaii/app/MainActivity.kt
@@ -1,10 +1,10 @@
 package org.gulaii.app
 
+import AppNavHost
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.navigation.compose.rememberNavController
-import org.gulaii.app.ui.navigation.AppNavHost
 import org.gulaii.app.ui.theme.GulaiiTheme
 
 class MainActivity : ComponentActivity() {
