+++ app/src/main/java/org/gulaii/app/ui/navigation/Screen.kt
@@ -1,0 +1,6 @@
+package org.gulaii.app.ui.navigation
+
+sealed class Screen(val route: String) {
+  object Initial : Screen("initial_screen")
+  object Onboarding : Screen("onboarding_screen")
+}
